[MS] Removed SettingsPage to use only SettingsModal
